feat(workspace): Expand test data and improve request name generation
This commit introduces a couple of changes to the 'TestNameGeneration' and 'RequestNameGeneration' classes.

The number of items included in the 'TestNameGeneration' test data has been increased from 3 to 6. This change will allow for more robust testing of the request name generation functionality.

In the 'RequestNameGenerator' class, several improvements have been made to the query function. The timeout for the OpenAI client has been increased to 25 seconds to account for possible network latency. The function now handles potential exceptions that could occur during the querying process, thus making the function more robust.

A new enum, 'GPTModel', was also added. This enum represents the different GPT models available in OpenAI and will make it easier to switch between different models in the future.

Overall, these changes lead to more robust and flexible code and improves the testing of the request name generation functionality.